Adding logging levels

+++ include/PDEParameters1D.h
@@ -8,7 +8,7 @@ public:
 
     PDEParameters1D(
         double max_timestep_length,
-        double nb_rannacher_steps,
+        size_t nb_rannacher_steps,
         double max_rannacher_step_length,
         size_t space_grid_size,
         double space_grid_nb_std_devs
@@ -17,14 +17,14 @@ public:
     virtual ~PDEParameters1D();
 
     double getMaxTimestepLength()      const;
-    double getNbRannacherSteps()       const;
+    size_t getNbRannacherSteps()       const;
     double getMaxRannacherStepLength() const;
     size_t getSpaceGridSize()          const;
     double getSpaceGridNbStdDevs()     const;
 
 protected:
     double m_maxTimestepLength;
-    double m_nbRannacherSteps;
+    size_t m_nbRannacherSteps;
     double m_maxRannacherStepLength;
     size_t m_spaceGridSize;
     double m_spaceGridNbStdDevs;
